Add python quantum circuit latex generator
This commit adds a python module for building quantum circuits in latex
with a python interface. Currently the latex circuit visualizer is
manually building the latex code by writing strings directly to a file
in a big series of nested loops. This gives no option to adjust spacing
or anything after the fact because once the iterators move on the string
is written out. This new module will give us a new interface for writing
the latex out using the pylatex library. It provides an interface to
build a latex circuit object which we can interact with, tweak as
necessary, and then write out the latex after everything is built.