Add compat shim for vqe benchmark on old versions of qiskit
When we're benchmarking old releases of qiskit the APIs used for running
the vqe benchmarks are different. To account for the differences this
commit adds an alternative path based on the version qiskit being used
and uses the old APIs for versions < 0.5.0. This should enable benchmark
comparisons between old and new releases.