Add custom build stage to asv config
In asv release 0.3.0 a new option, build_command, was added to the asv
conf.json file which enables running custom commands for the build
stage. This commit leverages this to build qiskit-terra using cmake.
This will enable running qiskit on earlier releases when we didn't
package a setup.py in repo, but instead relied on cmake to generate it.