Terminate cpp simulator on SIGINT and SIGTERM
This commit adds a signal handler that will terminate the subprocess
used to run the cpp local qasm simulator cleanly when the python side of
the code receives sigterm or sigint. Prior to this commit nothing would
tell the subprocess to terminate on these signals which caused the
signal to effectively be ignored while python waited for the cpp side to
exit. This address this and should make any locally running simulation
exit on a sigterm or sigint.

Fixes #939