Add constructor methods to build QuantumCircuits from QASM
This commit adds 2 new static methods to the QuantumCircuit class to
enable constructing a QuantumCircuit from a QASM string or file. This
gives users to option of building a QuantumCircuit object with a qasm
file simply by calling:

circuit = QuantumCircuit.from_qasm_file(path)

or

circuit = QuantumCircuit.from_qasm_str(qasm_str)

As part of this change the circuitbackend imports for Unroller() had to
be updated to be module imports. This was to avoid having a circular
import with the _quantumcircuit module (now that _quantumcircuit
leverages uses it).

Fixes #971