Add custom gate support to latex circuit drawer
This commit adds support to the latex circuit drawer backend to draw
custom gates. Currently when a circuit with a custom element is passed
into the latex circuit drawer it is silently ignored and not drawn. This
supports a custom gate with any number of qubits. This also has a bonus
effect if there was a standard gate which wasn't accounted for in the
latex drawer (we've had bugs like this in the past, see #1880) this will
also fix that for us, assuming the node name matches the expected gate
name (which isn't always the case). The only limitation with this is that
it doesn't support conditionals on the custom gates.

Fixes #1942