Reorganize QuantumCircuit to just access the ParameterTable directly
This commit updates all the parameter usage in QuantumCircuit to remove
any usage of self.parameters and instead access the ParameterTable
object directly. All the use of an intermediate set was unecessary and
caused unecessary slow downs. A dictionary (or in this case a mutable
mapping subclass) contains all the same performance characteristics of
a set for lookup already so there is no reason to convert it. The
changes made in the previous commits are left in place however, while
the performance bottleneck for parameterized circuit construction is
fixed by the changes in this commit the parameters() method of
QuantumCircuit is still a public api and those changes improve the
performance of that by not having to iterate over all elements every
time it's called by a user.

At the same time there were also unecessary set comprehension being used
which became the next bottleneck after these fixes. This commit removes
that and just accesses the ParameterTable directly. Which removes the
last scaling limit for large number of parameters and further speeds up
parameterized circuit creation.