Adds OperatorBase support in aux_operators setters (Qiskit/qiskit-aqua#1022)
* Adds OperatorBase support in aux_operators setters

aux_ops setters in vqe and numpy_eigen_solver call to_opflow on all operators,
whereas this conversion is only necessary for LegacyBaseOperator ---
BaseOperator does not need it and does not support it.

This commit adds a check in both setters, applying the conversion only
to instances of LegacyBaseOperator.

Fixes Qiskit/qiskit-aqua#1018

* fix vqe aux_operators setter

* fix aux_operators setter type hints

* Remove ListOp convert on aux_operators setter

Co-authored-by: Manoel Marques <manoel@us.ibm.com>
Co-authored-by: Manoel Marques <Manoel.Marques@ibm.com>