Cache decompositions in UnitarySynthesis and 1q optimization
This commit adds a caching dictionary to cache the output from unitary
synthesis for both the UnitarySynthesis pass and the
Optimize1qGatesDecomposition pass. The slowest portion of both of these
passes is synthesizing the unitary and currently we always do the work
everytime for every unitary, even if we've synthesized the circuit for
the unitary once before. This commit improves the performance of the
pass by caching the decompositions so we don't spend time synthesizing
the same unitaries more than once.