Simplify dynamical decoupling
This commit simplifies the dynamical decoupling pass. Instead of
iterating over the whole DAG and reconstructing the entire DAG from
scratch with dynamical decoupling sequences added where there are
delays. This switches the pass to only iterate over delays in the
DAG and substitutes the delays with the DD sequence.