Use backend_compat module's convert to target
This commit fixes an issue with the fake backend classes built on the
BackendV2 that used ECR or CZ (or any 2q gate besides CX) as gates. The
fundamental issue was the fake backend class had a hardcoded dict
mapping gate names to classes to use for generating a target. This
prevents the correct gate class from being reported when a backend uses
gates outside of that dict. To fix this issue this commit switches the
fake backends to use the convert_to_target function which is part of the
BackendConverterV2 class. That function provides the same functionality
as the one used internally by the fake provider except it is more
general as its designed to work with any backend. At the time of its
introduction this unification wasn't done because the fake backend's
version would be more efficient because it doesn't generate unused
intermediate object forms. But, a small loading time
performance regression against correct operation and deduplicating code
paths using the more general function is a better choice. In a future
commit we can deprecate and remove the fake backend version of
convert_to_target.

Fixes #10004
Fixes #9983