Prepare 1.0.0 release
This commit prepares the repo for the 1.0.0 release. It migrates all the
release notes to a dedicated directory to differentiate those notes from
future notes (this will need to be forward ported to the main branch),
updates the version numbers to indicate it is the final 1.0.0 release,
and updates the release notes to prepare them for publishing. Once this
merges this commit should be used to tag the 1.0.0 release.