Use SingletonGate for all parameterized non-controlled stdlib gates
This commit updates the parameterized non-controlled standard library
gates to leverage the SingletonGate class. The SingletonGate has
machinery to store a unique singleton instance for given constructor
arguments. The example at the time was to use the ctrl_state for
ControlledGate instances to enable using a singleton instance for open
control CXGates. This commit leverages this machinery to store a unique
singleton instance for each unique parameter value. This means whenever
there is a repeated parameter value used for a gate we only store a
single instance.

The potential caveat here is that the growth of the cache of singleton
instances could grow uncontrollably if there were never any exactly
repeated values (being floats most of the time this is a real risk).
There isn't a memory leak if all the unique instances are still retained
because we're not using any memory, but when any instances are deleted
their "singleton" will be retained and that could cause a memory leak
in scenarios when lots of unique parameter value gates are created and
deleted. To ameliorate this risk, this commit adds a new rust pyclass
LRUDict which implements an LRU cache internally and exposes itself as
a Python mapping. By default we only will retain the singleton instance
for the first 1000 parameter values, which should be ~500kbytes (if my
old estimates of ~500 bytes per gate object hold true).