Add parallel 1q euler decomposition interface
This commit adds a new inner rust function that performs 1q euler
decomposition in multiple threads to attempt to improve the performance
of the synthesis. The numerical component of the 1q synthesis has no
data dependency when run with multiple unitary matrices and we can
potentially better leverage multiple CPUs by performing the calculations
in parallel. We are still required to serially manipulate the DAGCircuit
though which will potentially limit the efficacy of this change. Longer
term if we move the core structural representation of DAGCircuit into
qiskit._accelerate we can potentially also parallelize the dag
manipulation (albeit with some locking).

This is the 1q component of #12007, we'll need to add similar interfaces
for TwoQubitBasisDecomposer, after #12004 is implemented, and
XXDecomposer, after #12005 is complete. The benefits of this are
expected to be much smaller compared to the 2q decomposers because the
computational effort in the numerical component of 1q euler
decomposition is very small.