Preallocate return Vecs for OneQubitGateSequence returns
This commit makes a small change to the euler_one_qubit_decomposer rust
module. It updates the circuit sequence generators to pre-allocate the
output Vecs to be the worst case size to reduce the number of
allocations performed. This might end up overallocating slightly but in
the worst case (for the psx generator) it's only overallocating by 4
elements which isn't very large and as the objects are only typically
temporary before we build a circuit (or dag) from the sequence.