Pre-synthesize SWAPs in SabreSwap
This commit makes a small tweak to the SabreSwap and SabreLayout passes
so that when a Target is present any inserted swap gates are added to
the circuit in the target's instructions. For example if a swap were to
be inserted on a qubit pair that only supported cx then instead of
inserting the swap gate 3 cx gates will be inserted. This avoids the
need for the translation stage to translate the gates in place which
saves a little work.