Guard against instruction params referencing the outer circuit
There were two tests that were putting the outer circuit an instruction
is a member of in that instructions parameters list. This is like making
a for loops body the outer circuit. In practice I don't know how this
works or what the expected use case is. But now that we've fixed
parameter handling for circuits this was causing failures. This commit
fixes the edge case by adding a custom path that processes the
parameters list in python instead of in rust.