Implement RGate in Rust according to established pattern
All the things are done to implement R(theta, phi) in Rust to the according to the pattern
set in the gates-in-rust PR. The "definition" is left as a todo again following the same
pattern.

Implementing this is rather clumsy and laborious. Doing the further gates would of
course be a bit easier. Some of this is due to interfaces beyond our control. We could
probably add a few conveniences that would make a difference. For example implementing
some conveniences for complex numbers.