Avoid Python op creation in commutative cancellation
This commit updates the commutative cancellation and commutation
analysis transpiler pass. It builds off of #12692 to adjust access
patterns in the python transpiler path to avoid eagerly creating a
Python space operation object. The goal of this PR is to mitigate the
performance regression on these passes introduced by the extra
conversion cost of #12459.

As part of this the commutation checker is rewritten in rust since all
that requires is gates in rust which we've had a representation of
since #12459 merged.